Introduce DestroySite

+++ app/services/destroy_user.rb
@@ -31,11 +31,13 @@ class DestroyUser
 
   def destroy_own_sites
     SiteMembership.where(user_id: user.id, role: 'owner').each do |membership|
+      next membership.destroy! unless membership.site
+
       if membership.site.users.count > 1
         promote_first_user_to_owner(membership)
         membership.destroy!
       else
-        membership.site.destroy!
+        DestroySite.new(membership.site).call
       end
     end
   end
